prussdrv: Remove unnecessary pthreads dependency
This removes the helper function for creating a pthread and the rest of
the infrastructure around it. Now the library can be linked in
environments that don't have pthreads. Also, the pthread helper code is
easily replicated in a user application and probably makes more sense
there anyway so that the user is more aware of the threading.

Signed-off-by: Frank Hunleth <fhunleth@troodon-software.com>