pasm:  Replace missing directory separator in `#include` for POSIX
Commit be231a82 fixed included files for Windows, but broke POSIX.
The POSIX `dirname()` removes a trailing `/`, but the Windows
equivalents do not.

Commit 731fab64 attempted to fix this by removing the trailing slash,
but broken pointer arithmetic ended up having no effect in almost
every case.

This commit attempts to fix the last fix, and then adds the `/` back
into the path at the correct place.  It also fixes cosmetic cases
where the included file is in the same directory, and the leading `./`
may be omitted.

Fixes #53